requirements.txt
Flask==2.0.2
Flask_Bcrypt==0.7.1
Flask_Login==0.5.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==0.15.1
Pillow==8.4.0
python-dotenv==0.19.2
SQLAlchemy==1.4.23
WTForms==2.3.3
